create kplot_widget
