+++ lase/gui/lidar_widget.py
@@ -4,7 +4,9 @@ from pyqtgraph.Qt import QtCore, QtGui
 import pyqtgraph as pg
 import numpy as np
 import time
-import datetime as datetime
+
+from kplot_widget import KPlotWidget
+from kplot_widget import TimeAxisItem
 
 class LidarWidget(QtGui.QWidget):
     def __init__(self, left_panel_layout):
@@ -58,7 +60,7 @@ class LidarWidget(QtGui.QWidget):
             
             self.plotWid.getPlotItem().getAxis('bottom').setLabel('Frequency (MHz)')
             self.plotWid.getPlotItem().getAxis('left').setLabel('PSD')
-        else:
+        else: # spectrum_plot
             self.is_velocity_plot = True
             self.velocity_plot_button.setText('Spectrum')
             
@@ -76,20 +78,5 @@ class LidarWidget(QtGui.QWidget):
         self.plotWid.getPlotItem().setMouseEnabled(x=False, y=True)
         self.plotWid.getViewBox().setMouseMode(self.plotWid.getViewBox().PanMode)
         self.plotWid.getPlotItem().enableAutoRange()
-            
-class KPlotWidget(pg.PlotWidget):
-    def __init__(self, *args, **kwargs):
-        super(KPlotWidget, self).__init__(*args, **kwargs)
-        
-        self.dataItem = pg.PlotDataItem(pen=(0,4), clear=True, _callSync='off')
-        self.addItem(self.dataItem)
-            
-# https://gist.github.com/friendzis/4e98ebe2cf29c0c2c232
-class TimeAxisItem(pg.AxisItem):
-    def __init__(self, *args, **kwargs):
-        super(TimeAxisItem, self).__init__(*args, **kwargs)
-
-    def tickStrings(self, values, scale, spacing):
-        return [datetime.datetime.fromtimestamp(value).strftime('%H:%M:%S') for value in values]
